bibsched_task: Adds deliver/doi-to-package data
* Changes the files bst_elsevier.py, bst_springer.py and bst_oxford.py
  to store the delivery and the mapping of a doi to a package name
  to a database.

* Adds bst_utils.py to unify the general procedure.
